+++ packages/iov-bns/src/types.ts
@@ -15,6 +15,7 @@ import {
   SwapOfferTransaction,
   TimestampTimeout,
 } from "@iov/bcp";
+import { isUint8Array } from "@iov/encoding";
 import { As } from "type-tagger";
 
 // config (those are not used outside of @iov/bns)
@@ -486,4 +487,12 @@ export function isBnsTx(transaction: LightTransaction): transaction is BnsTx {
   }
 
   return transaction.kind.startsWith("bns/");
+}
+
+export interface MultisignatureTx extends LightTransaction {
+  readonly multisig: readonly Uint8Array[];
+}
+
+export function isMultisignatureTx(transaction: LightTransaction): transaction is MultisignatureTx {
+  return Array.isArray((transaction as any).multisig) && (transaction as any).multisig.every(isUint8Array);
 }